Fix tvm verification
